+++ paypal/standard/forms.py
@@ -14,7 +14,8 @@ from django.utils.translation import ugettext_lazy as _
 
 from paypal.standard.conf import (
     DONATION_IMAGE, DONATION_SANDBOX_IMAGE, IMAGE, POSTBACK_ENDPOINT, SANDBOX_IMAGE, SANDBOX_POSTBACK_ENDPOINT,
-    SUBSCRIPTION_IMAGE, SUBSCRIPTION_SANDBOX_IMAGE
+    SUBSCRIPTION_IMAGE, SUBSCRIPTION_SANDBOX_IMAGE, PAYPAL_PRIVATE_CERT, PAYPAL_PUBLIC_CERT, PAYPAL_CERT,
+    PAYPAL_CERT_ID
 )
 from paypal.standard.widgets import ReservedValueHiddenInput, ValueHiddenInput
 
@@ -226,17 +227,20 @@ class PayPalEncryptedPaymentsForm(PayPalPaymentsForm):
 
     """
 
+    def __init__(self, private_cert=PAYPAL_PRIVATE_CERT, public_cert=PAYPAL_PUBLIC_CERT,
+            paypal_cert=PAYPAL_CERT, cert_id=PAYPAL_CERT_ID, *args, **kwargs):
+        super(PayPalEncryptedPaymentsForm, self).__init__(*args, **kwargs)
+        self.private_cert = private_cert
+        self.public_cert = public_cert
+        self.paypal_cert = paypal_cert
+        self.cert_id = cert_id
+
     def _encrypt(self):
         """Use your key thing to encrypt things."""
         from M2Crypto import BIO, SMIME, X509
-        # @@@ Could we move this to conf.py?
-        CERT = settings.PAYPAL_PRIVATE_CERT
-        PUB_CERT = settings.PAYPAL_PUBLIC_CERT
-        PAYPAL_CERT = settings.PAYPAL_CERT
-        CERT_ID = settings.PAYPAL_CERT_ID
 
         # Iterate through the fields and pull out the ones that have a value.
-        plaintext = 'cert_id=%s\n' % CERT_ID
+        plaintext = 'cert_id=%s\n' % self.cert_id
         for name, field in self.fields.items():
             value = None
             if name in self.initial:
@@ -252,9 +256,9 @@ class PayPalEncryptedPaymentsForm(PayPalPaymentsForm):
 
         # Begin crypto weirdness.
         s = SMIME.SMIME()
-        s.load_key_bio(BIO.openfile(CERT), BIO.openfile(PUB_CERT))
+        s.load_key_bio(BIO.openfile(self.private_cert), BIO.openfile(self.public_cert))
         p7 = s.sign(BIO.MemoryBuffer(plaintext), flags=SMIME.PKCS7_BINARY)
-        x509 = X509.load_cert_bio(BIO.openfile(PAYPAL_CERT))
+        x509 = X509.load_cert_bio(BIO.openfile(self.paypal_cert))
         sk = X509.X509_Stack()
         sk.push(x509)
         s.set_x509_stack(sk)