+++ paypal/standard/forms.py
@@ -73,6 +73,8 @@ class PayPalDateTimeField(forms.DateTimeField):
             # PST/PDT is 'US/Pacific'
             dt = timezone.pytz.timezone('US/Pacific').localize(
                 dt, is_dst=zone_part == 'PDT')
+            if not settings.USE_TZ:
+                dt = timezone.make_naive(dt, timezone=timezone.utc)
         return dt