+++ paypal/pro/models.py
@@ -9,6 +9,8 @@ from django.http import QueryDict
 from django.utils.functional import cached_property
 from django.utils.http import urlencode
 
+from paypal.utils import warn_untested
+
 try:
     from idmapper.models import SharedMemoryModel as Model
 except ImportError:
@@ -91,6 +93,7 @@ class PayPalNVP(Model):
         Returns a (MultiValueDict) dictionary containing all the parameters returned in the PayPal response.
         """
         # Undo the urlencode done in init
+        warn_untested()
         return QueryDict(self.response)
 
     def init(self, request, paypal_request, paypal_response):
@@ -98,6 +101,7 @@ class PayPalNVP(Model):
         if request is not None:
             self.ipaddress = request.META.get('REMOTE_ADDR', '').split(':')[0]
             if hasattr(request, "user") and request.user.is_authenticated():
+                warn_untested()
                 self.user = request.user
         else:
             self.ipaddress = ''
@@ -111,6 +115,7 @@ class PayPalNVP(Model):
         ack = paypal_response.get('ack', False)
         if ack != "Success":
             if ack == "SuccessWithWarning":
+                warn_untested()
                 self.flag_info = paypal_response.get('l_longmessage0', '')
             else:
                 self.set_flag(paypal_response.get('l_longmessage0', ''), paypal_response.get('l_errorcode', ''))
@@ -124,6 +129,7 @@ class PayPalNVP(Model):
 
     def process(self, request, item):
         """Do a direct payment."""
+        warn_untested()
         from paypal.pro.helpers import PayPalWPP
 
         wpp = PayPalWPP(request)