+++ paypal/pro/views.py
@@ -10,6 +10,7 @@ from django.template.response import TemplateResponse
 from paypal.pro.exceptions import PayPalFailure
 from paypal.pro.forms import ConfirmForm, PaymentForm
 from paypal.pro.helpers import PayPalWPP, express_endpoint_for_token
+from paypal.utils import warn_untested
 
 
 class PayPalPro(object):
@@ -137,6 +138,7 @@ class PayPalPro(object):
 
     def validate_payment_form(self):
         """Try to validate and then process the DirectPayment form."""
+        warn_untested()
         form = self.payment_form_cls(self.request.POST)
         if form.is_valid():
             success = form.process(self.request, self.item)
@@ -158,6 +160,7 @@ class PayPalPro(object):
         try:
             nvp_obj = wpp.setExpressCheckout(self.item)
         except PayPalFailure:
+            warn_untested()
             self.context['errors'] = self.errors['paypal']
             return self.render_payment_form()
         else:
@@ -168,6 +171,7 @@ class PayPalPro(object):
         Second step of ExpressCheckout. Display an order confirmation form which
         contains hidden fields with the token / PayerID from PayPal.
         """
+        warn_untested()
         initial = dict(token=self.request.GET['token'], PayerID=self.request.GET['PayerID'])
         self.context[self.form_context_name] = self.confirm_form_cls(initial=initial)
         return TemplateResponse(self.request, self.confirm_template, self.context)
@@ -184,6 +188,7 @@ class PayPalPro(object):
         # @@@ This check and call could be moved into PayPalWPP.
         try:
             if self.is_recurring():
+                warn_untested()
                 nvp = wpp.createRecurringPaymentsProfile(self.item)
             else:
                 nvp = wpp.doExpressCheckoutPayment(self.item)