+++ paypal/standard/forms.py
@@ -18,6 +18,7 @@ from paypal.standard.conf import (
     SUBSCRIPTION_SANDBOX_IMAGE
 )
 from paypal.standard.widgets import ReservedValueHiddenInput, ValueHiddenInput
+from paypal.utils import warn_untested
 
 log = logging.getLogger(__name__)
 
@@ -210,12 +211,15 @@ class PayPalPaymentsForm(forms.Form):
         }[self.test_mode(), self.button_type]
 
     def is_transaction(self):
+        warn_untested()
         return not self.is_subscription()
 
     def is_donation(self):
+        warn_untested()
         return self.button_type == self.DONATE
 
     def is_subscription(self):
+        warn_untested()
         return self.button_type == self.SUBSCRIBE
 
 
@@ -231,6 +235,7 @@ class PayPalEncryptedPaymentsForm(PayPalPaymentsForm):
 
     def __init__(self, private_cert=PAYPAL_PRIVATE_CERT, public_cert=PAYPAL_PUBLIC_CERT,
             paypal_cert=PAYPAL_CERT, cert_id=PAYPAL_CERT_ID, *args, **kwargs):
+        warn_untested()
         super(PayPalEncryptedPaymentsForm, self).__init__(*args, **kwargs)
         self.private_cert = private_cert
         self.public_cert = public_cert
@@ -239,6 +244,7 @@ class PayPalEncryptedPaymentsForm(PayPalPaymentsForm):
 
     def _encrypt(self):
         """Use your key thing to encrypt things."""
+        warn_untested()
         from M2Crypto import BIO, SMIME, X509
 
         # Iterate through the fields and pull out the ones that have a value.
@@ -273,6 +279,7 @@ class PayPalEncryptedPaymentsForm(PayPalPaymentsForm):
         return out.read()
 
     def as_p(self):
+        warn_untested()
         return mark_safe(u"""
 <input type="hidden" name="cmd" value="_s-xclick" />
 <input type="hidden" name="encrypted" value="%s" />
@@ -290,6 +297,7 @@ class PayPalSharedSecretEncryptedPaymentsForm(PayPalEncryptedPaymentsForm):
 
     def __init__(self, *args, **kwargs):
         "Make the secret from the form initial data and slip it into the form."
+        warn_untested()
         from paypal.standard.helpers import make_secret
 
         super(PayPalSharedSecretEncryptedPaymentsForm, self).__init__(*args, **kwargs)