+++ paypal/standard/helpers.py
@@ -5,8 +5,11 @@ import hashlib
 from django.conf import settings
 from django.utils.encoding import smart_str
 
+from paypal.utils import warn_untested
+
 
 def get_sha1_hexdigest(salt, raw_password):
+    warn_untested()
     return hashlib.sha1(smart_str(salt) + smart_str(raw_password)).hexdigest()
 
 
@@ -40,6 +43,7 @@ def make_secret(form_instance, secret_fields=None):
     selection of variables in params. Should only be used with SSL.
 
     """
+    warn_untested()
     # @@@ Moved here as temporary fix to avoid dependancy on auth.models.
     # @@@ amount is mc_gross on the IPN - where should mapping logic go?
     # @@@ amount / mc_gross is not nessecarily returned as it was sent - how to use it? 10.00 vs. 10.0
@@ -71,6 +75,7 @@ def check_secret(form_instance, secret):
     Used to verify IPN.
 
     """
+    warn_untested()
     # @@@ add invoice & custom
     # secret_fields = ['business', 'item_name']
     return make_secret(form_instance) == secret