+++ paypal/standard/models.py
@@ -10,6 +10,8 @@ from django.utils.functional import cached_property
 
 from paypal.standard.conf import POSTBACK_ENDPOINT, SANDBOX_POSTBACK_ENDPOINT
 from paypal.standard.helpers import check_secret, duplicate_txn_id
+from paypal.utils import warn_untested
+
 
 ST_PP_ACTIVE = 'Active'
 ST_PP_CANCELLED = 'Cancelled'
@@ -274,24 +276,31 @@ class PayPalStandardBase(Model):
         return len(self.recurring_payment_id) > 0
 
     def is_subscription(self):
+        warn_untested()
         return len(self.subscr_id) > 0
 
     def is_subscription_payment(self):
+        warn_untested()
         return self.txn_type == "subscr_payment"
 
     def is_subscription_failed(self):
+        warn_untested()
         return self.txn_type == "subscr_failed"
 
     def is_subscription_cancellation(self):
+        warn_untested()
         return self.txn_type == "subscr_cancel"
 
     def is_subscription_end_of_term(self):
+        warn_untested()
         return self.txn_type == "subscr_eot"
 
     def is_subscription_modified(self):
+        warn_untested()
         return self.txn_type == "subscr_modify"
 
     def is_subscription_signup(self):
+        warn_untested()
         return self.txn_type == "subscr_signup"
 
     def is_recurring_create(self):
@@ -310,18 +319,23 @@ class PayPalStandardBase(Model):
         return self.txn_type == "recurring_payment_failed"
 
     def is_recurring_suspended(self):
+        warn_untested()
         return self.txn_type == "recurring_payment_suspended"
 
     def is_recurring_suspended_due_to_max_failed_payment(self):
+        warn_untested()
         return self.txn_type == "recurring_payment_suspended_due_to_max_failed_payment"
 
     def is_billing_agreement(self):
+        warn_untested()
         return len(self.mp_id) > 0
 
     def is_billing_agreement_create(self):
+        warn_untested()
         return self.txn_type == "mp_signup"
 
     def is_billing_agreement_cancel(self):
+        warn_untested()
         return self.txn_type == "mp_cancel"
 
     def set_flag(self, info, code=None):
@@ -329,6 +343,7 @@ class PayPalStandardBase(Model):
         self.flag = True
         self.flag_info += info
         if code is not None:
+            warn_untested()
             self.flag_code = code
 
     def clear_flag(self):
@@ -365,6 +380,7 @@ class PayPalStandardBase(Model):
 
     def verify_secret(self, form_instance, secret):
         """Verifies an IPN payment over SSL using EWP."""
+        warn_untested()
         if not check_secret(form_instance, secret):
             self.set_flag("Invalid secret. (%s)") % secret
         self.save()