+++ paypal/standard/pdt/models.py
@@ -10,6 +10,7 @@ from six.moves.urllib.parse import unquote_plus
 
 from paypal.standard.conf import POSTBACK_ENDPOINT, SANDBOX_POSTBACK_ENDPOINT
 from paypal.standard.models import PayPalStandardBase
+from paypal.utils import warn_untested
 
 
 # ### Todo: Move this logic to conf.py:
@@ -46,10 +47,12 @@ class PayPalPDT(PayPalStandardBase):
         SUCCESS or FAILED.
 
         """
+        warn_untested()
         return requests.post(self.get_endpoint(),
                              data=dict(cmd="_notify-synch", at=IDENTITY_TOKEN, tx=self.tx)).content
 
     def get_endpoint(self):
+        warn_untested()
         if getattr(settings, 'PAYPAL_TEST', True):
             return SANDBOX_POSTBACK_ENDPOINT
         else:
@@ -67,6 +70,7 @@ class PayPalPDT(PayPalStandardBase):
                 self.st = unquoted_line
             else:
                 if self.st != "SUCCESS":
+                    warn_untested()
                     self.set_flag(line)
                     break
                 try: