+++ docs/release_notes.rst
@@ -2,8 +2,8 @@
  Release notes
 ===============
 
-Version 0.3.4 (under development)
----------------------------------
+Version 0.3.4
+-------------
 
 * Use multi certificates with PaypalEncryptedPaymentsForm
 * Fixed issue #166 - regression from 0.2.7 when using ``USE_TZ=False``